+++ sdk/clientcore/annotation-processor-test/src/main/java/io/clientcore/annotation/processor/test/MockHttpResponse.java
@@ -6,7 +6,7 @@ package io.clientcore.annotation.processor.test;
 import io.clientcore.core.http.models.HttpHeaders;
 import io.clientcore.core.http.models.HttpRequest;
 import io.clientcore.core.implementation.http.HttpResponse;
-import io.clientcore.core.serialization.JsonSerializer;
+import io.clientcore.core.serialization.json.JsonSerializer;
 import io.clientcore.core.models.binarydata.BinaryData;
 import io.clientcore.core.serialization.ObjectSerializer;
