+++ sdk/clientcore/core/src/main/java/io/clientcore/core/utils/CoreUtils.java
@@ -288,7 +288,7 @@ public final class CoreUtils {
      * @param typeArguments The type arguments.
      * @return The instance of {@link ParameterizedType}.
      */
-    public static ParameterizedType getParameterizedType(Type rawType, Type... typeArguments) {
+    public static ParameterizedType createParameterizedType(Type rawType, Type... typeArguments) {
         return new ParameterizedType() {
             @Override
             public Type[] getActualTypeArguments() {
@@ -371,33 +371,6 @@ public final class CoreUtils {
             default:
                 return String.join(delimiter, values);
         }
-    }
-
-    /**
-     * Create a parameterized type from a raw class and its type arguments.
-     *
-     * @param rawClass The raw class to construct the parameterized type.
-     * @param genericTypes The generic arguments.
-     *
-     * @return The parameterized type.
-     */
-    public static ParameterizedType createParameterizedType(Type rawClass, Type... genericTypes) {
-        return new ParameterizedType() {
-            @Override
-            public Type[] getActualTypeArguments() {
-                return genericTypes;
-            }
-
-            @Override
-            public Type getRawType() {
-                return rawClass;
-            }
-
-            @Override
-            public Type getOwnerType() {
-                return null;
-            }
-        };
     }
 
     private CoreUtils() {
