+++ sdk/clientcore/core/src/samples/java/io/clientcore/core/models/binarydata/BinaryDataJavaDocCodeSnippet.java
@@ -3,7 +3,7 @@
 
 package io.clientcore.core.models.binarydata;
 
-import io.clientcore.core.serialization.JsonSerializer;
+import io.clientcore.core.serialization.json.JsonSerializer;
 import io.clientcore.core.instrumentation.logging.ClientLogger;
 import io.clientcore.core.models.Person;
 import io.clientcore.core.serialization.ObjectSerializer;
